% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/insertCurlyBraces.R,
%   R/insertSquareBrackets.R
\name{insertSquareBrackets}
\alias{insertSquareBrackets}
\alias{insertSquareBrackets}
\alias{insertSquareBrackets}
\title{Adds an roxygen2 comment block}
\usage{
insertSquareBrackets()

insertSquareBrackets()

insertSquareBrackets()
}
\description{
Adds an roxygen2 comment block

Adds square brackets

Adds square brackets
}
